+++ assets/js/src/editor.js
@@ -1,0 +1,9 @@
+import { registerPlugin } from '@wordpress/plugins';
+import { default as PostAuthorIP } from './components/post-author-ip';
+
+registerPlugin(
+	'post-author-ip',
+	{
+		render: PostAuthorIP,
+	}
+);